Move `Domainic::Type::Constraint::WithAccessQualification`
This moves `Domainic::Type::Constraint::WithAccessQualification` to
`Domainic::Type::Constraint::Specification::WithAccessQualification`
